string table editor: created helpers.go
