+++ hswidget/stringtablewidget/widget.go
@@ -1,7 +1,6 @@
 package stringtablewidget
 
 import (
-	"sort"
 	"strconv"
 	"strings"
 
@@ -45,32 +44,7 @@ func (p *widget) Build() {
 func (p *widget) buildTableLayout() {
 	state := p.getState()
 
-	keys := make([]string, 0)
-
-	switch {
-	case state.numOnly:
-		for _, key := range state.keys {
-			if key[0] == '#' {
-				keys = append(keys, key)
-			} else {
-				// labels are sorted, so no-name (starting from # are on top)
-				break
-			}
-		}
-	case state.search != "":
-		for _, key := range state.keys {
-			k := strings.ToLower(key)
-			v := strings.ToLower(p.dict[key])
-
-			switch {
-			case strings.Contains(k, state.search),
-				strings.Contains(v, state.search):
-				keys = append(keys, key)
-			}
-		}
-	default:
-		keys = state.keys
-	}
+	keys := p.generateTableKeys()
 
 	rows := make([]*giu.RowWidget, len(keys)+1)
 
@@ -129,6 +103,37 @@ func (p *widget) buildTableLayout() {
 			giu.FastTable("").Border(true).Rows(rows),
 		}),
 	}.Build()
+}
+
+func (p *widget) generateTableKeys() (keys []string) {
+	state := p.getState()
+
+	switch {
+	case state.numOnly:
+		for _, key := range state.keys {
+			if key[0] == '#' {
+				keys = append(keys, key)
+			} else {
+				// labels are sorted, so no-name (starting from # are on top)
+				break
+			}
+		}
+	case state.search != "":
+		for _, key := range state.keys {
+			k := strings.ToLower(key)
+			v := strings.ToLower(p.dict[key])
+
+			switch {
+			case strings.Contains(k, state.search),
+				strings.Contains(v, state.search):
+				keys = append(keys, key)
+			}
+		}
+	default:
+		keys = state.keys
+	}
+
+	return
 }
 
 func (p *widget) buildAddEditLayout() {
@@ -200,47 +205,4 @@ func (p *widget) makeKeyField(id string) giu.Widget {
 		p.formatKey(&state.key)
 		p.updateValueText()
 	})
-}
-
-func (p *widget) calculateFirstFreeNoName() (firstFreeNoName int) {
-	state := p.getState()
-
-	ints := make([]int, 0)
-
-	for _, key := range state.keys {
-		if key[0] == '#' {
-			idx, err := strconv.Atoi(key[1:])
-			if err != nil {
-				continue
-			}
-
-			ints = append(ints, idx)
-		}
-	}
-
-	sort.Ints(ints)
-
-	for n, i := range ints {
-		if n != i {
-			firstFreeNoName = n
-			break
-		}
-	}
-
-	return
-}
-
-func (p *widget) updateValueText() {
-	state := p.getState()
-
-	str, found := p.dict[state.key]
-	if found {
-		state.value = str
-	} else {
-		state.value = ""
-	}
-}
-
-func (p *widget) formatKey(s *string) {
-	*s = strings.ReplaceAll(*s, " ", "_")
 }
